move destinations structures to separate file
