+++ redash/destinations.go
@@ -22,189 +22,6 @@ import (
 	// log "github.com/sirupsen/logrus"
 )
 
-// Type of Destination
-type Destination struct {
-	ID   int    `json:"id,omitempty"`
-	Name string `json:"name,omitempty"`
-	Type string `json:"type,omitempty"`
-	Icon string `json:"icon,omitempty"`
-}
-
-// DestinationType stucture
-type DestinationTypes struct {
-	Destination
-	ConfigurationSchema interface{}
-	//interface{} `json:"configuration_schema"`
-}
-
-type EmailDestination struct {
-	Destination
-	ConfgurationSchema struct {
-		Type       string `json:"type,omitempty"`
-		Properties struct {
-			Addresses struct {
-				Type string `json:"type,omitempty"`
-			} `json:"addresses"`
-			SubjectTemplate struct {
-				Type    string `json:"type,omitempty"`
-				Default string `json:"default,omitempty"`
-				Title   string `json:"title,omitempty"`
-			} `json:"subject_template"`
-		} `json:"properties"`
-		Required     []string `json:"required,omitempty"`
-		ExtraOptions []string `json:"extra_options,omitempty"`
-	} `json:"configuration_schema,omitempty"`
-}
-type SlackDestination struct {
-	Destination
-	ConfgurationSchema struct {
-		Type       string `json:"type"`
-		Properties struct {
-			URL struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"url,omitempty"`
-			Username struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"username,omitempty"`
-			IconEmoji struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"icon_emoji,omitempty"`
-			IconURL struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"icon_url,omitempty"`
-			Channel struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"channel,omitempty"`
-		} `json:"properties,omitempty"`
-		Secret []string `json:"secret,omitempty"`
-	} `json:"configuration_schema,omitempty"`
-}
-
-type WebhookDestination struct {
-	Destination
-	ConfgurationSchema struct {
-		Type       string `json:"type"`
-		Properties struct {
-			URL struct {
-				Type string `json:"type,omitempty"`
-			} `json:"url,omitempty"`
-			Username struct {
-				Type string `json:"type,omitempty"`
-			} `json:"username,omitempty"`
-			Password struct {
-				Type string `json:"type,omitempty"`
-			} `json:"password,omitempty"`
-		} `json:"properties,omitempty"`
-		Required []string `json:"required,omitempty"`
-		Secret   []string `json:"secret,omitempty"`
-	} `json:"configuration_schema,omitempty"`
-}
-
-type HipChatDestination struct {
-	Destination
-	ConfgurationSchema struct {
-		Type       string `json:"type"`
-		Properties struct {
-			URL struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"url,omitempty"`
-		} `json:"properties,omitempty"`
-		Secret   []string `json:"secret,omitempty"`
-		Required []string `json:"required,omitempty"`
-	} `json:"configuration_schema,omitempty"`
-	Deprecated bool `json:"deprecated,omitempty"`
-}
-type MattermostDestination struct {
-	Destination
-	ConfgurationSchema struct {
-		Type       string `json:"type"`
-		Properties struct {
-			URL struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"url,omitempty"`
-			Username struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"username,omitempty"`
-			IconURL struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"icon_url,omitempty"`
-			Channel struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"channel,omitempty"`
-		} `json:"properties,omitempty"`
-		Secret string `json:"secret,omitempty"`
-	} `json:"configuration_schema,omitempty"`
-}
-type ChatWorkDestination struct {
-	Destination
-	ConfgurationSchema struct {
-		Type       string `json:"type"`
-		Properties struct {
-			APIToken struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"api_token,omitempty"`
-			RoomID struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"room_id,omitempty"`
-			MessageTemplate struct {
-				Type    string `json:"type,omitempty"`
-				Default string `json:"default,omitempty"`
-				Title   string `json:"title,omitempty"`
-			} `json:"message_template,omitempty"`
-		} `json:"properties,omitempty"`
-		Secret   []string `json:"secret,omitempty"`
-		Required []string `json:"required,omitempty"`
-	} `json:"configuration_schema,omitempty"`
-}
-type PagerDutyDestination struct {
-	Destination
-	ConfgurationSchema struct {
-		Type       string `json:"type"`
-		Properties struct {
-			IntegrationKey struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"integration_key,omitempty"`
-			Description struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"description,omitempty"`
-		} `json:"properties,omitempty"`
-		Secret   []string `json:"secret,omitempty"`
-		Required []string `json:"required,omitempty"`
-	} `json:"configuration_schema,omitempty"`
-}
-type HangoutsChatDestination struct {
-	Destination
-	ConfgurationSchema struct {
-		Type       string `json:"type,omitempty"`
-		Properties struct {
-			URL struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"url"`
-			IconURL struct {
-				Type  string `json:"type,omitempty"`
-				Title string `json:"title,omitempty"`
-			} `json:"icon_url"`
-		} `json:"properties"`
-		Secret   []string `json:"secret,omitempty"`
-		Required []string `json:"required,omitempty"`
-	} `json:"configuration_schema,omitempty"`
-}
-
 func ParseDestinationType(payload []byte) (destination interface{}, err error) {
 	dst := &Destination{}
 	err = json.Unmarshal(payload, dst)
